Refactor and add PrivateRoute
