+++ project/src/components/layout/footer/footer.tsx
@@ -1,9 +1,12 @@
+import { Link } from "react-router-dom";
+import { ERoute } from "../../../types/enums/route.enum";
+
 export default function Footer(): JSX.Element {
   return (
     <footer className="footer">
-      <a className="footer__logo-link" href="main.html">
+      <Link className="footer__logo-link" to={ERoute.MAIN}>
         <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33"/>
-      </a>
+      </Link>
     </footer>
   );
 }
