+++ project/src/pages/login-page/login-page.tsx
@@ -1,3 +1,6 @@
+import { Link } from "react-router-dom";
+import { ERoute } from "../../types/enums/route.enum";
+
 export default function LoginPage(): JSX.Element {
   return (
     <div className="page page--gray page--login">
@@ -5,7 +8,7 @@ export default function LoginPage(): JSX.Element {
         <div className="container">
           <div className="header__wrapper">
             <div className="header__left">
-              <a className="header__logo-link" href="main.html">
+              <Link className="header__logo-link" to={ERoute.MAIN}>
                 <img
                   className="header__logo"
                   src="img/logo.svg"
@@ -13,7 +16,7 @@ export default function LoginPage(): JSX.Element {
                   width="81"
                   height="41"
                 />
-              </a>
+              </Link>
             </div>
           </div>
         </div>
